app-client/requirements.txt
fastapi
uvicorn
pyjwt
cryptography
requests
